Created new classes for scrapers
